Parsing of incomming messages are segfaulting

+++ dtndht/dht.h
@@ -39,8 +39,8 @@ int dht_init(int s, int s6, const unsigned char *id, const unsigned char *v);
 int dht_insert_node(const unsigned char *id, struct sockaddr *sa, int salen);
 int dht_ping_node(struct sockaddr *sa, int salen);
 int dht_ping_dtn_node(struct sockaddr *sa, int salen);
-int dht_add_dtn_eid(const char *eid, int eidlen, enum dtn_dht_lookup_type type);
-int dht_remove_dtn_eid(const char *eid, int eidlen);
+int dht_add_dtn_eid(const char *eid, size_t eidlen, enum dtn_dht_lookup_type type);
+int dht_remove_dtn_eid(const char *eid, size_t eidlen);
 int dht_periodic(const void *buf, size_t buflen, const struct sockaddr *from,
 		int fromlen, time_t *tosleep, dht_callback *callback, void *closure,
 		const struct dtn_dht_context * ctx);
